refactor: delete event by id updated on services and repository to delete only events of the authenticated user

+++ src/respositories/EventsRepositories/EventsRepository.ts
@@ -24,9 +24,9 @@ export default class EventsRepository {
     const event = await Event.findById(eventId)
     return event
   }
-  static async deleteEventById(eventId: string) {
+  static async deleteEventById(eventId: string, userId: string) {
     const eventToDelete = await Event.findById(eventId)
-    await Event.findByIdAndDelete(eventId)
+    await Event.findByIdAndDelete({ eventId, userId })
     return eventToDelete
   }
 }
